Migrate code to the proper file to account for refactoring
